backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
pydantic>=2.9.0
pydantic-settings>=2.6.0
qdrant-client>=1.12.0
google-genai>=1.0.0
Pillow>=11.0.0
python-dotenv>=1.0.0
